pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sv-vault"
version = "2.2.0"
description = "분산 시크릿 관리 — AES-256-GCM + Argon2id + Shamir's Secret Sharing"
readme = "SPEC.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "MeshPOP", email = "fools@fools.ai"},
]
keywords = [
    "vault", "secret", "encryption", "shamir", "distributed",
    "aes-256-gcm", "argon2id", "password-manager",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "cryptography>=41.0",
    "argon2-cffi>=23.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
sv = "cli:main"
vault-mcp = "vault_mcp_server:main"

[project.urls]
Homepage = "https://github.com/meshpop/vault"

[tool.setuptools]
py-modules = [
    "vault_mcp_server",
    "engine",
    "keymanager",
    "transport",
    "config",
    "backup",
    "vault",
    "cli",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py39"
